requirements.txt
asgiref==3.8.1
bcrypt==4.2.0
Django==5.1.2
djangorestframework==3.15.2
greenlet==3.1.1
psycopg2==2.9.9
PyJWT==2.9.0
pytz==2024.2
SQLAlchemy==2.0.35
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
